refactor: onboarding history fields are whole counts (int, no decimals)
- All three fields int? end-to-end (domain, DTOs, contracts, handlers,
  Dapper, EF, API, UI, tests)
- DB column type: integer (nullable) in migration/designer/snapshot
- UI validates whole numbers only (no decimals), numeric keypad,
  swagger integer/int32

+++ src/Application/Features/Employees/GetUserEmployeeForEditQuery.cs
@@ -17,7 +17,7 @@ public record GetUserEmployeeForEditQueryResponse(
     string PhoneNumber,
     string? JobTitle,
     Region Region,
-    decimal? LeaveUsedInCurrentYear,
-    decimal? NetWorkedDaysBeforeCurrentMonth,
-    decimal? CarriedOverLeaveFromPreviousYear,
+    int? LeaveUsedInCurrentYear,
+    int? NetWorkedDaysBeforeCurrentMonth,
+    int? CarriedOverLeaveFromPreviousYear,
     List<EmployeeBankAccountDto> BankAccounts);
