+++ src/Web.Api/Controllers/Employees/Contracts/GetEmployeeForEditRequest.cs
@@ -12,7 +12,7 @@ public record GetEmployeeForEditResponse(
     string PhoneNumber,
     string? JobTitle,
     Region Region,
-    decimal? LeaveUsedInCurrentYear,
-    decimal? NetWorkedDaysBeforeCurrentMonth,
-    decimal? CarriedOverLeaveFromPreviousYear,
+    int? LeaveUsedInCurrentYear,
+    int? NetWorkedDaysBeforeCurrentMonth,
+    int? CarriedOverLeaveFromPreviousYear,
     List<EmployeeBankAccountResponse> BankAccounts);
